sketch out the static types a bit more
