+++ packages/@molt/command/src/index_.ts
@@ -1,1 +1,5 @@
-export { initializeViaParameter as parameter, initializeViaParameters as parameters } from './builder.js'
+export {
+  initializeViaParameter as parameter,
+  initializeViaParameters as parameters,
+  initializeViaParametersExclusive as parametersExclusive,
+} from './Builder/initializers.js'
